+++ bmad/agents/Agent/AiDeveloper/aidev.py
@@ -5,12 +5,13 @@ Implementeert en integreert AI/ML-functionaliteit. Output in prompt templates, c
 """
 
 import argparse
-import sys
+import sys, os
+sys.path.append(os.path.abspath(os.path.join(os.path.dirname(__file__), "../../../..")))
 import textwrap
 
-from bmad.agents.core.message_bus import publish, subscribe
-from bmad.agents.core.supabase_context import save_context, get_context
-from bmad.agents.core.llm_client import ask_openai
+from bmad.agents.core.communication.message_bus import publish, subscribe
+from bmad.agents.core.data.supabase_context import save_context, get_context
+from bmad.agents.core.ai.llm_client import ask_openai
 
 
 class AiDeveloperAgent:
