+++ bmad/agents/Agent/Architect/architect.py
@@ -1,10 +1,12 @@
 import logging
 import time
 from pathlib import Path
+import sys, os
+sys.path.append(os.path.abspath(os.path.join(os.path.dirname(__file__), "../../../..")))
 
-from bmad.agents.core.message_bus import publish, subscribe
-from bmad.agents.core.supabase_context import save_context, get_context
-from bmad.agents.core.llm_client import ask_openai
+from bmad.agents.core.communication.message_bus import publish, subscribe
+from bmad.agents.core.data.supabase_context import save_context, get_context
+from bmad.agents.core.ai.llm_client import ask_openai
 
 logging.basicConfig(level=logging.INFO, format="[%(levelname)s] %(message)s")
