+++ bmad/agents/Agent/DevOpsInfra/devopsinfra.py
@@ -1,5 +1,10 @@
-from bmad.agents.core.llm_client import ask_openai
-from bmad.agents.core.message_bus import subscribe, publish
+import sys, os
+sys.path.append(os.path.abspath(os.path.join(os.path.dirname(__file__), "../../../..")))
+from bmad.agents.core.communication.message_bus import publish, subscribe
+from bmad.agents.core.data.supabase_context import save_context, get_context
+from bmad.agents.core.ai.llm_client import ask_openai
+from bmad.agents.core.ai.confidence_scoring import confidence_scoring
+from integrations.slack.slack_notify import send_slack_message
 import logging
 import time
