+++ bmad/agents/Agent/DocumentationAgent/documentationagent.py
@@ -1,9 +1,13 @@
 import glob
 import logging
-from bmad.agents.core.slack_notify import send_slack_message
-from bmad.agents.core.llm_client import ask_openai
-from bmad.agents.core.message_bus import subscribe
-from bmad.agents.core.figma_client import FigmaClient
+import sys, os
+sys.path.append(os.path.abspath(os.path.join(os.path.dirname(__file__), "../../../..")))
+from bmad.agents.core.communication.message_bus import publish, subscribe
+from bmad.agents.core.data.supabase_context import save_context, get_context
+from bmad.agents.core.ai.llm_client import ask_openai
+from bmad.agents.core.ai.confidence_scoring import confidence_scoring
+from integrations.slack.slack_notify import send_slack_message
+from integrations.figma.figma_client import FigmaClient
 import json
 from typing import Dict, List, Optional
