+++ bmad/agents/Agent/FullstackDeveloper/fullstackdeveloper.py
@@ -11,9 +11,11 @@ import logging
 import time
 from dotenv import load_dotenv
 load_dotenv()
+import sys, os
+sys.path.append(os.path.abspath(os.path.join(os.path.dirname(__file__), "../../../..")))
 
-from bmad.agents.core.message_bus import publish, subscribe
-from bmad.agents.core.supabase_context import save_context, get_context
+from bmad.agents.core.communication.message_bus import publish, subscribe
+from bmad.agents.core.data.supabase_context import save_context, get_context
 
 
 class FullstackDeveloperAgent:
