+++ bmad/agents/Agent/ProductOwner/product_owner.py
@@ -5,10 +5,12 @@ Product Owner Agent voor BMAD
 import argparse
 import logging
 import time
-from bmad.agents.core.message_bus import publish, subscribe
-from bmad.agents.core.supabase_context import save_context, get_context
-from bmad.agents.core.llm_client import ask_openai_with_confidence as ask_openai
-from bmad.agents.core.confidence_scoring import confidence_scoring, create_review_request, format_confidence_message
+import sys, os
+sys.path.append(os.path.abspath(os.path.join(os.path.dirname(__file__), "../../../..")))
+from bmad.agents.core.communication.message_bus import publish, subscribe
+from bmad.agents.core.data.supabase_context import save_context, get_context
+from bmad.agents.core.ai.llm_client import ask_openai_with_confidence as ask_openai
+from bmad.agents.core.ai.confidence_scoring import confidence_scoring, create_review_request, format_confidence_message
 from bmad.projects.project_manager import project_manager
 from dotenv import load_dotenv
 load_dotenv()
